+++ r2-streamer-swiftTests/Parser/EPUB/Services/EPUBPositionsServiceTests.swift
@@ -1,12 +1,7 @@
 //
-//  EPUBPositionsServiceTests.swift
-//  r2-streamer-swift
-//
-//  Created by Mickaël Menu on 01/06/2020.
-//
 //  Copyright 2020 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import XCTest
@@ -334,11 +329,11 @@ class EPUBPositionsServiceTests: XCTestCase {
         ])
     }
     
-    func testUsesOriginalLengthWhenAvailable() {
+    func testUsesArchiveEntryLengthWhenAvailable() {
         let service = makeService(
             layout: .reflowable,
             readingOrder: [
-                (60, Link(href: "chap1", properties: makeProperties(originalLength: 20))),
+                (60, Link(href: "chap1", properties: makeProperties(archiveEntryLength: 20))),
                 (60, Link(href: "chap2"))
             ],
             reflowablePositionLength: 50
@@ -390,15 +385,15 @@ func makeService(layout: EPUBLayout? = nil, readingOrder: [(UInt64, Link)], refl
     )
 }
 
-private func makeProperties(layout: EPUBLayout? = nil, originalLength: Int? = nil) -> Properties {
+private func makeProperties(layout: EPUBLayout? = nil, archiveEntryLength: UInt64? = nil) -> Properties {
     var props: [String: Any] = [:]
     if let layout = layout {
         props["layout"] = layout.rawValue
     }
-    if let originalLength = originalLength {
-        props["encrypted"] = [
-            "algorithm": "algo",
-            "originalLength": originalLength
+    if let archiveEntryLength = archiveEntryLength {
+        props["archive"] = [
+            "entryLength": archiveEntryLength as NSNumber,
+            "isEntryCompressed": true
         ]
     }
     return Properties(props)